Move CrateContext into it's own file
